travis-ci is gonna yell at me, but I'm too tired to fix this right now
Realized that arithmetic instructions are supposed to affect flags, and wrote
some code in the cpu.dec_r8 method to try it. The test currently fails, either
because it works wrong or because I'm inaccurately predicting what the
resulting flags should be. Either way, my brain is fried and I'll look at it
tomorrow.